+++ jupyterhub/common.py
@@ -1,53 +1,155 @@
+import json
+import logging
 import os
 import string
 import sys
-import json
+import time
 
+import requests
 from tapipy.tapis import Tapis
+
+logger = logging.getLogger("JupyterHub")
 
 INSTANCE = os.environ.get("INSTANCE")
 TENANT = os.environ.get("TENANT")
+RESTRICTED_ID = os.environ.get("RESTRICTED_ID", "66657")
+RESTRICTED_LABEL = os.environ.get("RESTRICTED_LABEL", "hetdex")
 tapis_service_token = os.environ.get("TAPIS_SERVICE_TOKEN")
-base_url = os.environ.get("AGAVE_BASE_URL", "https://api.tacc.utexas.edu")
+portals_service_token = os.environ.get("PORTALS_SERVICE_TOKEN")
+portals_base_url = os.environ.get("PORTALS_BASE_URL", "https://portals.tapis.io")
 tapis_base_url = os.environ.get("TAPIS_BASE_URL", "https://tacc.tapis.io")
-v2_token_url = os.environ.get("V2_TOKEN_URL", "https://tacc.develop.tapis.io/v3/oauth2/v2/token")
+meta_base_url = os.environ.get("META_BASE_URL", "https://tacc.tapis.io")
 database = os.environ.get("TAPIS_DATABASE")
 collection = os.environ.get("TAPIS_COLLECTION")
+
 
 if not tapis_service_token:
     raise Exception("Missing TAPIS_SERVICE_TOKEN configuration.")
 
 
-def get_config_metadata_name():
+def get_metadata(t, q):
+    try:
+        response = json.loads(t.meta.listDocuments(db=database, collection=collection, filter=json.dumps(q)))
+    except Exception as e:
+        logger.error(f"Unable to get metadata, error: {e}")
+        return None
+    return response
+
+
+def get_config_metadata_name(restricted):
     """Return name of config metadata"""
-    return f"config.{TENANT}.{INSTANCE}.jhub"
-
-
-def get_tenant_configs():
-    """Retrive tenant config from metadata"""
-    t = Tapis(base_url=tapis_base_url, jwt=tapis_service_token)
-    q = {"name": get_config_metadata_name()}
-    print(f"tenant query: {q}")
-    metadata = json.loads(
-        t.meta.listDocuments(db=database, collection=collection, filter=json.dumps(q))
-    )[0]["value"]
-    return metadata
-
-
-def get_user_configs(username):
-    """Retrieve any groups user belongs to"""
-    t = Tapis(base_url=tapis_base_url, jwt=tapis_service_token)
-    q = {"value.user": username, "value.tenant": TENANT, "value.instance": INSTANCE}
-    print(f"user query: {q}")
-    metadata = json.loads(
-        t.meta.listDocuments(db=database, collection=collection, filter=json.dumps(q))
+    return (
+        f"config.{TENANT}.{INSTANCE}.jhub"
+        if not restricted
+        else f"config.{TENANT}.{INSTANCE}.restricted.jhub"
     )
+
+
+def get_tenant_configs(restricted=False, retry=True):
+    """Retrive tenant config from metadata"""
+    t = Tapis(base_url=meta_base_url, jwt=tapis_service_token)
+    q = {"name": get_config_metadata_name(restricted)}
+    metadata = get_metadata(t, q)
+    if not metadata:
+        if retry:
+            time.sleep(1)
+            return get_tenant_configs(restricted, False)
+        return None
+    logger.error(f"Loaded tenant config: {metadata}")
+    return metadata[0]["value"]
+
+
+def get_user_configs(username, retry=True):
+    """Retrieve any groups user belongs to"""
+    t = Tapis(base_url=meta_base_url, jwt=tapis_service_token)
+    q = {"value.user": username, "value.tenant": TENANT, "value.instance": INSTANCE}
+    metadata = get_metadata(t, q)
+    if not metadata:
+        if retry:
+            time.sleep(1)
+            return get_user_configs(username, False)
     return metadata
 
 
-def safe_string(
-    to_escape, safe=None, escape_char="-"
+def refresh_access_token(refresh_token, username):
+    logger.info(f"Refreshing access token for user {username}")
+    try:
+        data = {
+            "refresh_token": refresh_token,
+        }
+        res = requests.put("https://tacc.tapis.io/v3/tokens", json=data)
+        logger.debug(f"Token refresh response: {res}")
+        resp_data = res.json()
+        logger.debug(f"Token refresh data: {resp_data}")
+        new_access_token = resp_data["result"]["access_token"]["access_token"]
+        refresh_token = resp_data["result"]["refresh_token"]["refresh_token"]
+
+        expires_in = resp_data["result"]["access_token"]["expires_in"]
+        expires_at = resp_data["result"]["access_token"]["expires_at"]
+        return {
+            "access_token": new_access_token,
+            "refresh_token": refresh_token,
+            "created_at": time.time(),
+            "expires_in": expires_in,
+            "expires_at": expires_at,
+        }
+    except Exception as e:
+        logger.error(f"Unable to refresh access token for {username}, error: {e}")
+
+
+def get_user_token_dir(username):
+    return os.path.join("/tapis/jupyter/tokens", INSTANCE, TENANT, username)
+
+
+def save_token(
+    access_token, refresh_token, username, created_at, expires_in, expires_at
 ):
+    try:
+        configs = get_tenant_configs()
+        tenant_id = configs.get("tapis_tenant_id")
+        # tapipy file
+        d = [
+            {
+                "token": access_token,
+                "refresh_token": refresh_token,
+                "tenant_id": tenant_id,
+                "api_key": configs.get("tapis_client_id"),
+                "api_secret": configs.get("tapis_client_secret"),
+                "api_server": "{}".format(configs.get("tapis_base_url").rstrip("/")),
+                "verify": eval(configs.get("oauth_validate_cert")),
+            }
+        ]
+        with open(os.path.join(get_user_token_dir(username), ".tapipy"), "w") as f:
+            json.dump(d, f)
+        # logger.info(
+        #     f"Saved tapipy cache file to {os.path.join(get_user_token_dir(username), '.tapipy')}"
+        # )
+        logger.debug(f"tapipy cache file data: {d}")
+
+        # cli file
+        d = {
+            "tenantid": tenant_id,
+            "baseurl": "{}".format(configs.get("tapis_base_url").rstrip("/")),
+            "devurl": "",
+            "apikey": configs.get("tapis_client_id"),
+            "username": username,
+            "access_token": access_token,
+            "refresh_token": refresh_token,
+            "created_at": str(int(created_at)),
+            "apisecret": configs.get("tapis_client_secret"),
+            "expires_in": str(expires_in),
+            "expires_at": str(expires_at),
+        }
+        with open(os.path.join(get_user_token_dir(username), "current"), "w") as f:
+            json.dump(d, f)
+        # logger.info(
+        #     f"Saved CLI cache file to {os.path.join(get_user_token_dir(username), 'current')}"
+        # )
+    except Exception as e:
+        logger.error(f"Unable to save CLI cache file for {username}, error: {e}")
+
+
+def safe_string(to_escape, safe=None, escape_char="-"):
     """Escape a string so that it only contains characters in a safe set.
     Characters outside the safe list will be escaped with _%x_,
     where %x is the hex value of the character.
@@ -67,7 +169,6 @@ if sys.version_info >= (3,):
 
     def _ord(byte):
         return byte
-
 
 else:
     _ord = ord