+++ jupyterhub/spawner_hooks.py
@@ -1,22 +1,28 @@
 import ast
-import humanfriendly
 import json
 import os
 import re
-import requests
+import time
 
-from tornado import web
-from ldap3 import Server, Connection, SAFE_SYNC
-from agavepy.agave import Agave
+import certifi
+import humanfriendly
+import jwt
+import requests
+import urllib3
 from jupyterhub.common import (
-    TENANT,
     INSTANCE,
-    base_url,
-    v2_token_url,
+    RESTRICTED_ID,
+    RESTRICTED_LABEL,
+    TENANT,
     get_tenant_configs,
-    safe_string,
     get_user_configs,
+    refresh_access_token,
+    safe_string,
+    save_token,
+    tapis_base_url,
 )
+from ldap3 import SAFE_SYNC, Connection, Server
+from tornado import web
 
 # TAS configuration:
 # base URL for TAS API.
@@ -32,22 +38,55 @@ def hook(spawner):
     spawner.log.info("👽 user configs 👽 {}".format(spawner.user_configs))
     spawner.log.info("😱 user options (from form) 😱 {}".format(spawner.user_options))
 
-    get_agave_access_data(spawner)
+    # Add call to TAS to get allocation data for user
+    tas_data = get_user_projects(spawner)
+
+    # Check response from TAS to check for any allocation
+    allowed = is_user_allowed(spawner, tas_data)
+    if not allowed:
+        raise web.HTTPError(403)
+
+    # Check if user only has restricted allocation
+    restricted = is_user_restricted(spawner, tas_data)
+    spawner.log.info(f"Restricted? {restricted}")
+
+    # If user has allocation: keep as is
+    # spawner.configs = get_tenant_configs()
+    # If user only has specific "restricted" allocation:
+    # set spawner.configs to restricted metadata group
+    if restricted:
+        spawner.configs = get_tenant_configs(restricted)
+        spawner.log.info(f"spawner configs: {spawner.configs}")
+        spawner.user_configs = get_user_configs(spawner.user.name)
+        spawner.log.info(f"spawner user configs: {spawner.configs}")
+
+    # Check if access token is valid
+    get_tapis_access_data(spawner)
     spawner.log.info(
         "access token: {}, refresh token: {}, url: {}".format(
             spawner.access_token, spawner.refresh_token, spawner.url
         )
     )
-    # check if access token is valid
-    
-    get_tas_data(spawner)
 
-    spawner.uid = int(spawner.configs.get("uid", spawner.tas_uid))
-    spawner.gid = int(spawner.configs.get("gid", spawner.tas_gid))
+    # If in training instance, default to 100 uid/gid
+    if "training" not in tapis_base_url:
+        get_tas_data(spawner)
+        spawner.uid = spawner.tas_uid
+        spawner.gid = spawner.tas_gid
+    else:
+        spawner.uid = 100
+        spawner.gid = 100
 
+    # Retrieve all of the configs and merge them together
     spawner.extra_pod_config = spawner.configs.get("extra_pod_config", {})
     spawner.extra_container_config = spawner.configs.get("extra_container_config", {})
-    
+
+    for user_conf in spawner.user_configs:
+        if "extra_pod_config" in user_conf:
+            merge_configs(
+                user_conf["extra_pod_config"], spawner.extra_pod_config
+            )
+
     if (
         len(spawner.configs.get("images")) == 1 and not spawner.hpc_available
     ):  # only 1 image option, so we skipped the form
@@ -63,7 +102,7 @@ def hook(spawner):
             for image in item["value"]["images"]:
                 spawner.log.info(f"Image: {image}")
                 image_options.append(image)
-        user_options = spawner.user_options        
+        user_options = spawner.user_options
         spawner.log.info(f"User options: {user_options}")
         image = ast.literal_eval(spawner.user_options["image"][0])
         spawner.log.info(f"Image: {image}")
@@ -74,7 +113,10 @@ def hook(spawner):
                 )
             )
             allowed_options = next(
-                option for option in image_options if option["name"] == image["name"] and option["display_name"] == image["display_name"]
+                option
+                for option in image_options
+                if option["name"] == image["name"]
+                and option["display_name"] == image["display_name"]
             )
             if spawner.user_options.get("hpc"):
                 if not eval(allowed_options.get("hpc_available", "False")):
@@ -93,6 +135,8 @@ def hook(spawner):
             raise web.HTTPError(403)
 
         spawner.image = image["name"]
+        spawner.log.info(image)
+        spawner.log.info(spawner.extra_pod_config)
         if image.get("extra_pod_config"):
             merge_configs(image["extra_pod_config"], spawner.extra_pod_config)
         if image.get("extra_container_config"):
@@ -116,7 +160,8 @@ def hook(spawner):
         )
         spawner.mem_limit = max(mem_limits, key=mem_limits.get)
         spawner.cpu_limit = float(max(cpu_limits))
-        # Set the guarantees really low because when None or 0,it sets a resource request for an amount equal to the limit
+        # Set the guarantees really low because when None or 0,
+        # it sets a resource request for an amount equal to the limit
         spawner.mem_guarantee = ".001K"
         spawner.cpu_guarantee = float(0.001)
         spawner.environment = {
@@ -127,29 +172,87 @@ def hook(spawner):
             "SCINCO_JUPYTERHUB_IMAGE": spawner.image,
         }
     get_mounts(spawner)
-    get_projects(spawner)
 
 
 def merge_configs(x, y):
     merged_pod_config = {**x, **y}
     for key, value in merged_pod_config.items():
         if key in x and key in y:
-            merged_pod_config[key].update(x[key])
+            if type(x[key]) is list:
+                for z in x[key]:
+                    merged_pod_config[key].append(z)
+            else:
+                merged_pod_config[key].update(x[key])
+
+
+def get_user_projects(spawner):
+    """
+    Retrieve user projects from TAS
+    """
+    try:
+        user = spawner.user.name
+        http_headers = urllib3.make_headers(basic_auth=f"{TAS_ROLE_ACCT}:{TAS_ROLE_PASS}")
+        pool_manager = urllib3.PoolManager(
+            cert_reqs="CERT_REQUIRED",
+            ca_certs=certifi.where(),
+            retries=False,
+            headers=http_headers,
+        )
+        response = pool_manager.request("GET", f"{TAS_URL_BASE}/projects/username/{user}")
+        spawner.log.info(f"{TAS_URL_BASE}/projects/username/{user}")
+        json_response = json.loads(response.data.decode("utf-8"))
+        spawner.log.info(f"TAS Projects for {user}: {json_response}")
+        return json_response
+    except Exception:
+        return {}
+
+
+def is_user_allowed(spawner, tas_data):
+    """
+    A user is allowed if they have any allocation
+    """
+    user = spawner.user.name
+    spawner.log.info(f"Check if user has any allocation: {user}")
+    return bool(tas_data.get("result"))
+
+
+def is_user_restricted(spawner, tas_data):
+    """
+    If the user is only in the restricted HETDEX project,
+    they are given the restricted user configs.
+    """
+    user = spawner.user.name
+    spawner.log.info(f"Check tas for restricted project for user: {user}")
+    if len(tas_data["result"]) == 1:
+        item_id = tas_data["result"][0].get("id")
+        if item_id is not None and str(item_id) == RESTRICTED_ID:
+            spawner.log.info(f"Found restricted project for user: {user}")
+            spawner.extra_labels = {"restrictedProject": RESTRICTED_LABEL}
+            return True
+
+    return False
+
 
 async def get_notebook_options(spawner):
-    spawner.configs = get_tenant_configs()
+    tas_data = get_user_projects(spawner)
+    restricted = is_user_restricted(spawner, tas_data)
+    spawner.configs = get_tenant_configs(restricted=restricted)
     spawner.log.info(f"spawner configs: {spawner.configs}")
     spawner.user_configs = get_user_configs(spawner.user.name)
-    spawner.log.info(f"spawner user configs: {spawner.configs}")
+    spawner.log.info(f"spawner user configs: {spawner.user_configs}")
 
     image_options = spawner.configs.get("images")
 
-    for item in spawner.user_configs:
-        for image in item["value"].get("images"):
-            if image not in image_options:
-                image_options += [image]
-            if eval(image.get("hpc_available", "False")):
-                spawner.hpc_available = True
+    if spawner.user_configs:
+        for item in spawner.user_configs:
+            if isinstance(item, dict) and "value" in item:
+                images = item["value"].get("images")
+                if isinstance(images, list):
+                    for image in images:
+                        if image not in image_options:
+                            image_options += [image]
+                        if eval(image.get("hpc_available", "False")):
+                            spawner.hpc_available = True
 
     if not hasattr(
         spawner, "hpc_available"
@@ -209,33 +312,28 @@ async def get_notebook_options(spawner):
         select_images = '<select id="image" name="image" size="10" onchange="{}"> {} </select>'.format(
             js, options
         )
-        spawner.log.info(select_images)
         return "{}{}{}".format(select_images, image_description, hpc)
+
 
 async def parse_form_data(formdata, spawner):
     spawner.log.info(f"FORM DATA: {formdata}")
     return formdata
 
 
-def get_agave_access_data(spawner):
+def get_tapis_access_data(spawner):
     """
-    Returns the access token and base URL cached in the agavepy file
-    :return:
+    Returns the access token and base URL cached in the tapipy file
     """
     # TODO figure out naming conventions that can follow k8 rules
     # k8 names must consist of lower case alphanumeric characters, '-' or '.',
     # and must start and end with an alphanumeric character
     # do all tenant names follow that? usernames?
-    token_file = os.path.join(get_user_token_dir(spawner.user.name), ".agpy")
+    token_file = os.path.join(get_user_token_dir(spawner.user.name), ".tapipy")
     spawner.log.info(
-        "spawner looking for token file: {} for user: {}".format(
-            token_file, spawner.user.name
-        )
+        f"spawner looking for token file: {token_file} for user: {spawner.user.name}"
     )
     if not os.path.exists(token_file):
-        spawner.log.warning(
-            "spawner did not find a token file at {}".format(token_file)
-        )
+        spawner.log.warning(f"spawner did not find a token file at {token_file}")
         return None
     try:
         data = json.load(open(token_file))
@@ -245,11 +343,46 @@ def get_agave_access_data(spawner):
 
     try:
         spawner.access_token = data[0]["token"]
-        spawner.log.info("Setting token: {}".format(spawner.access_token))
-        spawner.refresh_token = data[0]["refresh_token"]
-        spawner.log.info("Setting refresh token: {}".format(spawner.refresh_token))
+        try:
+            decoded_data = jwt.decode(
+                data[0]["token"], options={"verify_signature": False}
+            )
+        except Exception as e:
+            print(f"Error decoding access token: {e}")
+
+        refresh_data = None
+        if "exp" in decoded_data and decoded_data["exp"] < time.time():
+            spawner.log.info(
+                f"{spawner.user.name} has expired access token, attempting to refresh"
+            )
+            refresh_data = refresh_access_token(
+                data[0]["refresh_token"], spawner.user.name
+            )
+            spawner.log.info(f"Data retrieved from refreshing: {refresh_data}")
+
+        if refresh_data:
+            spawner.log.info(
+                f"Refreshed access token for: {spawner.user.name}, attempting to save and update tapipy files"
+            )
+            save_token(
+                refresh_data["access_token"],
+                refresh_data["refresh_token"],
+                spawner.user.name,
+                refresh_data["created_at"],
+                refresh_data["expires_in"],
+                refresh_data["expires_at"],
+            )
+            spawner.access_token = refresh_data["access_token"]
+            # spawner.log.info(f"Setting token: {spawner.access_token}")
+            spawner.refresh_token = refresh_data["refresh_token"]
+            # spawner.log.info(f"Setting refresh token: {spawner.refresh_token}")
+        else:
+            # spawner.log.info(f"Setting token: {spawner.access_token}")
+            spawner.refresh_token = data[0]["refresh_token"]
+            # spawner.log.info(f"Setting refresh token: {spawner.refresh_token}")
+
         spawner.url = data[0]["api_server"]
-        spawner.log.info("Setting url: {}".format(spawner.url))
+        # spawner.log.info(f"Setting url: {spawner.url}")
 
     except (TypeError, KeyError):
         spawner.log.warning(
@@ -307,29 +440,35 @@ def get_tas_data(spawner):
             )
         )
         return
-    
+
     gids = []
 
     try:
-        server = Server('ldaps://ldap.tacc.utexas.edu:636')
-        conn = Connection(server, 'uid=ldapbind,ou=People,dc=tacc,dc=utexas,dc=edu', LDAP_PASS, client_strategy=SAFE_SYNC, auto_bind=True)
-        status, result, response, _ = conn.search('ou=Groups,dc=tacc,dc=utexas,dc=edu', f'(uniqueMember=uid={spawner.user.name},ou=People,dc=tacc,dc=utexas,dc=edu)')
+        server = Server("ldaps://ldap.tacc.utexas.edu:636")
+        conn = Connection(
+            server,
+            "uid=ldapbind,ou=People,dc=tacc,dc=utexas,dc=edu",
+            LDAP_PASS,
+            client_strategy=SAFE_SYNC,
+            auto_bind=True,
+        )
+        status, result, response, _ = conn.search(
+            "ou=Groups,dc=tacc,dc=utexas,dc=edu",
+            f"(uniqueMember=uid={spawner.user.name},ou=People,dc=tacc,dc=utexas,dc=edu)",
+        )
         for entry in response:
-            data = entry['dn'].split(',')
-            cn = data[0].split('=')
+            data = entry["dn"].split(",")
+            cn = data[0].split("=")
             group = cn[1]
-            temp_gid = group.split('-')[1]
+            temp_gid = group.split("-")[1]
             try:
                 gid = int(temp_gid)
                 gids.append(gid)
-            except Exception as e:
+            except Exception:
                 continue
     except Exception as e:
-        spawner.log.error(
-            "Did not get gid's from ldap. rsp: {}"
-            .format(e)
-        )
-    
+        spawner.log.error("Did not get gid's from ldap. rsp: {}".format(e))
+
     if gids:
         spawner.supplemental_gids = gids
 
@@ -348,19 +487,15 @@ def get_tas_data(spawner):
 
 
 def get_user_token_dir(username):
-    return os.path.join("/agave/jupyter/tokens", INSTANCE, TENANT, username)
+    return os.path.join("/tapis/jupyter/tokens", INSTANCE, TENANT, username)
 
 
 def get_mounts(spawner):
     safe_username = safe_string(spawner.user.name).lower()
     safe_tenant = safe_string(TENANT).lower()
     safe_instance = safe_string(INSTANCE).lower()
-    agpy_safe_name = "{}-{}-{}-jhub-agpy".format(
-        safe_username, safe_tenant, safe_instance
-    )
-    current_safe_name = "{}-{}-{}-jhub-current".format(
-        safe_username, safe_tenant, safe_instance
-    )
+    tapipy_safe_name = f"{safe_username}-{safe_tenant}-{safe_instance}-jhub-tapipy"
+    current_safe_name = f"{safe_username}-{safe_tenant}-{safe_instance}-jhub-current"
 
     spawner.init_containers = [
         {
@@ -369,62 +504,62 @@ def get_mounts(spawner):
             "command": [
                 "/bin/sh",
                 "-c",
-                "cp -r /agave_data/.agpy /agave_data_rw/.agpy && cp -r /agave_data/current /agave_data_rw/current && ls -lah /agave_data_rw && cat /agave_data_rw/current/current && chmod -R 777 /agave_data_rw && ls -lah /agave_data_rw",
+                "cp -r /tapis_data/.tapipy /tapis_data_rw/.tapipy && cp -r /tapis_data/current /tapis_data_rw/current && ls -lah /tapis_data_rw && cat /tapis_data_rw/current/current && chmod -R 777 /tapis_data_rw && ls -lah /tapis_data_rw",
             ],
             "volumeMounts": [
                 {
-                    "mountPath": "/agave_data/.agpy",
-                    "name": "{}-configmap".format(agpy_safe_name),
-                    "subPath": ".agpy",
+                    "mountPath": "/tapis_data/.tapipy",
+                    "name": f"{tapipy_safe_name}-configmap",
+                    "subPath": ".tapipy",
                 },
                 {
-                    "mountPath": "/agave_data/current",
-                    "name": "{}-configmap".format(current_safe_name),
+                    "mountPath": "/tapis_data/current",
+                    "name": f"{current_safe_name}-configmap",
                     "subPath": "current",
                 },
                 {
-                    "mountPath": "/agave_data_rw/.agpy",
-                    "name": agpy_safe_name,
-                    "subPath": ".agpy",
+                    "mountPath": "/tapis_data_rw/.tapipy",
+                    "name": tapipy_safe_name,
+                    "subPath": ".tapipy",
                 },
                 {
-                    "mountPath": "/agave_data_rw/current",
+                    "mountPath": "/tapis_data_rw/current",
                     "name": current_safe_name,
                     "subPath": "current",
-                }
+                },
             ],
         }
     ]
 
     spawner.volumes = [
         {
-            "name": "{}-configmap".format(agpy_safe_name),
-            "configMap": {"name": agpy_safe_name, "defaultMode": 0o0777},
+            "name": f"{tapipy_safe_name}-configmap",
+            "configMap": {"name": tapipy_safe_name, "defaultMode": 0o0777},
         },
         {
-            "name": "{}-configmap".format(current_safe_name),
+            "name": f"{current_safe_name}-configmap",
             "configMap": {"name": current_safe_name, "defaultMode": 0o0777},
         },
         {
-            "name": agpy_safe_name,
+            "name": tapipy_safe_name,
             "emptyDir": {},
         },
         {
             "name": current_safe_name,
             "emptyDir": {},
-        }
+        },
     ]
     spawner.volume_mounts = [
         {
-            "mountPath": "/etc/.agpy",
-            "name": agpy_safe_name,
-            "subPath": ".agpy/.agpy",
+            "mountPath": "/etc/.tapipy",
+            "name": tapipy_safe_name,
+            "subPath": ".tapipy/.tapipy",
         },
         {
-            "mountPath": "/home/jupyter/.agave",
+            "mountPath": "/home/jovyan/.tapis-token",
             "name": current_safe_name,
             "subPath": "current",
-        }
+        },
     ]
     volume_mounts = spawner.configs.get("volume_mounts")
 
@@ -449,7 +584,7 @@ def get_mounts(spawner):
             # volume names must consist of lower case alphanumeric characters or '-',
             # and must start and end with an alphanumeric character (e.g. 'my-name',  or '123-abc',
             # regex used for validation is '[a-z0-9]([-a-z0-9]*[a-z0-9])?')
-            if item["mountPath"][-1] == '/':
+            if item["mountPath"][-1] == "/":
                 item["mountPath"] = item["mountPath"][:-1]
             vol_name = re.sub(
                 r"([^a-z0-9-\s]+?)", "", item["mountPath"].split("/")[-1].lower()
@@ -460,10 +595,10 @@ def get_mounts(spawner):
                 vol["server"] = item["server"]
 
             if item["path"] == "/work2/{tas_homedir}":
-                spawner.log.info(spawner.init_gid)
+                # TODO: remove this check
                 if spawner.init_gid == 0:
                     continue
-            
+
             spawner.volumes.append({"name": vol_name, item["type"]: vol})
 
             spawner.volume_mounts.append(
@@ -471,106 +606,3 @@ def get_mounts(spawner):
             )
         spawner.log.info("volumes: {}".format(spawner.volumes))
         spawner.log.info("volume_mounts: {}".format(spawner.volume_mounts))
-
-
-def get_projects(spawner):
-    if not spawner.access_token:
-        spawner.log.info("no access_token")
-        return None
-    #url = "{}/projects/v2/".format(spawner.url)
-    projects_url = f"{base_url}/projects/v2"
-    
-    # use spawner.access_token to generate v2 token
-    # call v3 to v2 token endpoint
-    # tacc.develop.tapis.io/v3/oauth2/v2/token
-    spawner.log.error("getting projects")
-    try:
-        token_url = v2_token_url
-        headers = {
-            'x-tapis-token': spawner.access_token
-        }
-        rsp = requests.post(token_url, headers=headers)
-        rsp.raise_for_status()
-        access_token = rsp.json()['access_token']
-    except Exception as e:
-        spawner.log.error(f"Unable to generate v2 token; error: {e}; response: {rsp}")
-        return None
-
-    # with v2 token, send request to projects url
-    try:
-        ag = Agave(api_server=base_url, token=access_token)
-        rsp = ag.geturl(projects_url)
-        rsp.raise_for_status()
-        data = rsp.json()
-    except Exception as e:
-        spawner.log.warn(f"Did not get data from /projects. Exception: {e}")
-        spawner.log.warn(f"Full response from service: {rsp}")
-        spawner.log.warn(f"url used: {projects_url}")
-        return None
-
-    projects = data.get("mounts")
-    spawner.network_storage = spawner.configs.get('network_storage')
-
-    try:
-        spawner.log.info("Found {} projects".format(len(projects)))
-    except TypeError:
-        spawner.log.error("Projects data has no length.")
-        spawner.log.info(f"response: {rsp}, data: {data}")
-        return None
-
-    for p in projects:
-        mountPath = p.get('mountPath')
-        path = p.get('path')
-        pems = p.get('pems')
-        readOnly = False if pems == 'rw' else True
-        source = ""
-        if path.split('/')[1] == 'work':
-            source = "work"
-            path = path.replace('work', 'work2', 1)
-        if path.split('/')[1] == 'corral-repl':
-            source = "corral"
-            path = path.replace('corral-repl', 'corral/main', 1)
-        # check if corral -- then use nfs, else, use hostPath
-        if source == "corral":
-            name = 'project-{}'.format(safe_string(mountPath).lower())
-            if len(name) > 63:
-                name = name[:62] + 'e'
-            spawner.volumes.append(
-                {
-                    'name': name,
-                    "nfs": {
-                        "server": spawner.network_storage,
-                        "path": path,
-                        "readOnly": readOnly,
-                    },
-                }
-            )
-
-            spawner.volume_mounts.append(
-                {
-                    "mountPath": f"/home/jovyan/projects{mountPath}",
-                    'name': name,
-                }
-            )
-        if source == "work":
-            name = 'project-{}'.format(safe_string(mountPath).lower())
-            if len(name) > 63:
-                name = name[:62] + 'e'
-            spawner.volumes.append(
-                {
-                    'name': name,
-                    "hostPath": {
-                        "path": path,
-                        "readOnly": readOnly,
-                    },
-                }
-            )
-
-            spawner.volume_mounts.append(
-                {
-                    "mountPath": f"/home/jovyan/projects{mountPath}",
-                    'name': name,
-                }
-            )
-    spawner.log.info(spawner.volumes)
-    spawner.log.info(spawner.volume_mounts)