+++ CHANGELOG.md
@@ -1,9 +1,27 @@
 # Changelog
 All notable changes to this project will be documented in this file
 
+## [3.1.6] (May 8, 2026)
+### 🚀 Added
+- Updated metadata calls to be more robust
+### 🚀 Modified
+- Update to allocation checks
+
+## [3.1.3] (2025)
+### 🚀 Added
+- Checks for training accounts
+- Check for expired token, and refresh automatically
+- Assign restricted settings to users on restricted allocation
+- Block users that have no allocation
+### 🚀 Modified
+- Updated to JupyterHub v3
+### 🚀 Deleted
+- Removed agavepy and all references to it
+
 ## [2.0.0] (2021)
 ### 🚀 Added
 - JupyterHub now uses v3 metadata!
+
 ## [1.5.0] (2021-11-02)
 ### 🚀 Added
 - v3 authentication for JupyterHub
@@ -11,5 +29,6 @@ All notable changes to this project will be documented in this file
 - Initial commit to repo
 
 
+[3.1.3]: https://github.com/scinco-project/jhub/releases/tag/v3.1.3
 [2.0.0]: https://github.com/scinco-project/jhub/releases/tag/v2.0.0
 [1.5.0]: https://github.com/scinco-project/jhub/releases/tag/v1.5.0